requirements.txt
slackclient==2.9.3
slackeventsapi==2.2.1
